+++ Dungeon Jumak/Assets/00. Scripts/03. UI/CommonUI/Profile/ProfilePageSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d31b77515389157458af172ddbdd0965
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
