+++ Dungeon Jumak/Assets/00. Scripts/99. Data/02. UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b8c3143a38b319e4183c523e7bd5a591
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
