+++ Dungeon Jumak/Assets/00. Scripts/99. Data/02. UI/RatingData.cs
@@ -1,0 +1,18 @@
+// System
+using System;
+
+/// <summary>
+/// 등급과 관련된 데이터를 저장하는 클래스입니다.
+/// </summary>
+[Serializable]
+public class RatingData
+{
+    private int rating;
+    private int interiorScore;
+    private bool isPushEnabled = true;
+    private bool isNightPushEnabled = true;
+
+    #region Property - Get / Set
+
+    #endregion
+}
