+++ Dungeon Jumak/Assets/00. Scripts/99. Data/02. UI/RatingData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0013c9458de9c1b448338d2e694d7060
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
