+++ Dungeon Jumak/Assets/Resources/Prefabs/UI/PopUp/ProfilePopUp.prefab
@@ -675,6 +675,7 @@ GameObject:
   - component: {fileID: 3383550925921203960}
   - component: {fileID: 3383550925921203961}
   - component: {fileID: 1478281391532147187}
+  - component: {fileID: 6346660579642506434}
   m_Layer: 5
   m_Name: ProfilePopUp
   m_TagString: Untagged
@@ -778,7 +779,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 29b66c39505a5ff4cbde510c90aa7d68, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  ratingIcon:
+--- !u!114 &6346660579642506434
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3383550925921203966}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d31b77515389157458af172ddbdd0965, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ratingIcons:
   - {fileID: 292772859502925083}
   - {fileID: 8891716363903810703}
   - {fileID: 5772936246166017760}
